refactor: merge separate error modules with utils
