+++ demo/midgard-node/src/workers/utils/common.ts
